refactor: extract mapcontainer for common map
